requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.30.0
crispy-bootstrap5==0.6
cryptography==37.0.4
Django==4.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-rest-knox==4.2.0
djangorestframework==3.13.1
fontawesomefree==6.1.1
gunicorn==20.1.0
idna==3.4
Pillow==9.2.0
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stripe==5.0.0
urllib3==1.26.12
